backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.10
pydantic-settings==2.5.2
python-jose==3.4.0
bcrypt==4.1.3
alembic==1.13.3
email-validator==2.3.0
celery==5.4.0
redis==5.1.1

